requirements.txt
fastapi
uvicorn
python-jose
passlib
pydantic
python-dotenv
httpx
beautifulsoup4
cachetools
nltk
numpy
tensorflow
opencv-python
SpeechRecognition
pydub
moviepy
requests
motor
bson
certifi
google-generativeai
